fix filters
